chore(cli): async await test that expects rejected promise (#4)
Fixing this warning in CI

```
stderr | src/commands/deploy/__tests__/baremetal.test.js > deployTasks > throws an error if there is not enough available space on the server and freeSpaceRequired is not configured
Promise returned by `expect(actual).rejects.toThrowError(expected)` was not awaited. Vitest currently auto-awaits hanging assertions at the end of the test, but this will cause the test to fail in Vitest 3. Please remember to await the assertion.
    at /home/runner/work/graphql/graphql/packages/cli/src/commands/deploy/__tests__/baremetal.test.js:705:40
```